Revert "did a bunch of things"
