+++ .env
@@ -1,3 +1,3 @@
 DB_NAME='book_db'
-DB_PASSWORD='Maddie187!'
+DB_PASSWORD=''
 DB_USER='root'
