--- package.json
@@ -1,242 +1,0 @@
-{
-  "name": "bookworm",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "dependencies": {
-    "abbrev": "^1.1.1",
-    "accepts": "^1.3.8",
-    "acorn": "^7.4.1",
-    "acorn-jsx": "^5.3.2",
-    "agent-base": "^6.0.2",
-    "ajv": "^6.12.6",
-    "ansi-colors": "^4.1.3",
-    "ansi-regex": "^5.0.1",
-    "ansi-styles": "^4.3.0",
-    "aproba": "^2.0.0",
-    "are-we-there-yet": "^2.0.0",
-    "argparse": "^1.0.10",
-    "array-flatten": "^1.1.1",
-    "astral-regex": "^2.0.0",
-    "asynckit": "^0.4.0",
-    "axios": "^1.4.0",
-    "balanced-match": "^1.0.2",
-    "bcrypt": "^5.1.0",
-    "body-parser": "^1.20.1",
-    "brace-expansion": "^1.1.11",
-    "bytes": "^3.1.2",
-    "call-bind": "^1.0.2",
-    "callsites": "^3.1.0",
-    "chalk": "^4.1.2",
-    "chownr": "^2.0.0",
-    "color-convert": "^2.0.1",
-    "color-name": "^1.1.4",
-    "color-support": "^1.1.3",
-    "combined-stream": "^1.0.8",
-    "concat-map": "^0.0.1",
-    "connect-session-sequelize": "^7.1.7",
-    "console-control-strings": "^1.1.0",
-    "content-disposition": "^0.5.4",
-    "content-type": "^1.0.5",
-    "cookie": "^0.5.0",
-    "cookie-signature": "^1.0.6",
-    "cross-spawn": "^7.0.3",
-    "debug": "^2.6.9",
-    "deep-is": "^0.1.4",
-    "delayed-stream": "^1.0.0",
-    "delegates": "^1.0.0",
-    "denque": "^2.1.0",
-    "depd": "^2.0.0",
-    "destroy": "^1.2.0",
-    "detect-libc": "^2.0.1",
-    "doctrine": "^3.0.0",
-    "dotenv": "^8.6.0",
-    "dottie": "^2.0.3",
-    "ee-first": "^1.1.1",
-    "emoji-regex": "^8.0.0",
-    "encodeurl": "^1.0.2",
-    "enquirer": "^2.3.6",
-    "escape-html": "^1.0.3",
-    "escape-string-regexp": "^4.0.0",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^6.15.0",
-    "eslint-scope": "^5.1.1",
-    "eslint-utils": "^2.1.0",
-    "eslint-visitor-keys": "^2.1.0",
-    "espree": "^7.3.1",
-    "esprima": "^4.0.1",
-    "esquery": "^1.5.0",
-    "esrecurse": "^4.3.0",
-    "estraverse": "^4.3.0",
-    "esutils": "^2.0.3",
-    "etag": "^1.8.1",
-    "express": "^4.18.2",
-    "express-handlebars": "^5.3.5",
-    "express-session": "^1.17.3",
-    "fast-deep-equal": "^3.1.3",
-    "fast-json-stable-stringify": "^2.1.0",
-    "fast-levenshtein": "^2.0.6",
-    "file-entry-cache": "^6.0.1",
-    "finalhandler": "^1.2.0",
-    "flat-cache": "^3.0.4",
-    "flatted": "^3.2.7",
-    "follow-redirects": "^1.15.2",
-    "form-data": "^4.0.0",
-    "forwarded": "^0.2.0",
-    "fresh": "^0.5.2",
-    "fs-minipass": "^2.1.0",
-    "fs.realpath": "^1.0.0",
-    "function-bind": "^1.1.1",
-    "functional-red-black-tree": "^1.0.1",
-    "gauge": "^3.0.2",
-    "generate-function": "^2.3.1",
-    "get-intrinsic": "^1.2.1",
-    "get-stdin": "^6.0.0",
-    "glob": "^7.2.3",
-    "glob-parent": "^5.1.2",
-    "globals": "^13.20.0",
-    "graceful-fs": "^4.2.11",
-    "handlebars": "^4.7.7",
-    "has": "^1.0.3",
-    "has-flag": "^4.0.0",
-    "has-proto": "^1.0.1",
-    "has-symbols": "^1.0.3",
-    "has-unicode": "^2.0.1",
-    "http-errors": "^2.0.0",
-    "https-proxy-agent": "^5.0.1",
-    "iconv-lite": "^0.4.24",
-    "ignore": "^4.0.6",
-    "import-fresh": "^3.3.0",
-    "imurmurhash": "^0.1.4",
-    "inflection": "^1.13.4",
-    "inflight": "^1.0.6",
-    "inherits": "^2.0.4",
-    "ipaddr.js": "^1.9.1",
-    "is-extglob": "^2.1.1",
-    "is-fullwidth-code-point": "^3.0.0",
-    "is-glob": "^4.0.3",
-    "is-property": "^1.0.2",
-    "isexe": "^2.0.0",
-    "js-tokens": "^4.0.0",
-    "js-yaml": "^3.14.1",
-    "json-schema-traverse": "^0.4.1",
-    "json-stable-stringify-without-jsonify": "^1.0.1",
-    "levn": "^0.4.1",
-    "lodash": "^4.17.21",
-    "lodash.merge": "^4.6.2",
-    "lodash.truncate": "^4.4.2",
-    "long": "^4.0.0",
-    "lru-cache": "^6.0.0",
-    "make-dir": "^3.1.0",
-    "media-typer": "^0.3.0",
-    "merge-descriptors": "^1.0.1",
-    "methods": "^1.1.2",
-    "mime": "^1.6.0",
-    "mime-db": "^1.52.0",
-    "mime-types": "^2.1.35",
-    "minimatch": "^3.1.2",
-    "minimist": "^1.2.8",
-    "minipass": "^5.0.0",
-    "minizlib": "^2.1.2",
-    "mkdirp": "^1.0.4",
-    "moment": "^2.29.4",
-    "moment-timezone": "^0.5.43",
-    "ms": "^2.0.0",
-    "mysql2": "^2.3.3",
-    "named-placeholders": "^1.1.3",
-    "natural-compare": "^1.4.0",
-    "negotiator": "^0.6.3",
-    "neo-async": "^2.6.2",
-    "node-addon-api": "^5.1.0",
-    "node-fetch": "^2.6.11",
-    "nopt": "^5.0.0",
-    "npmlog": "^5.0.1",
-    "object-assign": "^4.1.1",
-    "object-inspect": "^1.12.3",
-    "on-finished": "^2.4.1",
-    "on-headers": "^1.0.2",
-    "once": "^1.4.0",
-    "optionator": "^0.9.1",
-    "parent-module": "^1.0.1",
-    "parseurl": "^1.3.3",
-    "path-is-absolute": "^1.0.1",
-    "path-key": "^3.1.1",
-    "path-to-regexp": "^0.1.7",
-    "pg-connection-string": "^2.6.0",
-    "prelude-ls": "^1.2.1",
-    "prettier": "^2.8.8",
-    "progress": "^2.0.3",
-    "proxy-addr": "^2.0.7",
-    "proxy-from-env": "^1.1.0",
-    "punycode": "^2.3.0",
-    "qs": "^6.11.0",
-    "random-bytes": "^1.0.0",
-    "range-parser": "^1.2.1",
-    "raw-body": "^2.5.1",
-    "readable-stream": "^3.6.2",
-    "regexpp": "^3.2.0",
-    "require-from-string": "^2.0.2",
-    "resolve-from": "^4.0.0",
-    "retry-as-promised": "^7.0.4",
-    "rimraf": "^3.0.2",
-    "safe-buffer": "^5.2.1",
-    "safer-buffer": "^2.1.2",
-    "semver": "^7.5.1",
-    "send": "^0.18.0",
-    "seq-queue": "^0.0.5",
-    "sequelize": "^6.31.1",
-    "sequelize-pool": "^7.1.0",
-    "serve-static": "^1.15.0",
-    "set-blocking": "^2.0.0",
-    "setprototypeof": "^1.2.0",
-    "shebang-command": "^2.0.0",
-    "shebang-regex": "^3.0.0",
-    "side-channel": "^1.0.4",
-    "signal-exit": "^3.0.7",
-    "slice-ansi": "^4.0.0",
-    "source-map": "^0.6.1",
-    "sprintf-js": "^1.0.3",
-    "sqlstring": "^2.3.3",
-    "statuses": "^2.0.1",
-    "string_decoder": "^1.3.0",
-    "string-width": "^4.2.3",
-    "strip-ansi": "^6.0.1",
-    "strip-json-comments": "^3.1.1",
-    "supports-color": "^7.2.0",
-    "table": "^6.8.1",
-    "tar": "^6.1.15",
-    "text-table": "^0.2.0",
-    "toidentifier": "^1.0.1",
-    "toposort-class": "^1.0.1",
-    "tr46": "^0.0.3",
-    "type-check": "^0.4.0",
-    "type-fest": "^0.20.2",
-    "type-is": "^1.6.18",
-    "uglify-js": "^3.17.4",
-    "uid-safe": "^2.1.5",
-    "unpipe": "^1.0.0",
-    "uri-js": "^4.4.1",
-    "util-deprecate": "^1.0.2",
-    "utils-merge": "^1.0.1",
-    "uuid": "^8.3.2",
-    "v8-compile-cache": "^2.3.0",
-    "validator": "^13.9.0",
-    "vary": "^1.1.2",
-    "webidl-conversions": "^3.0.1",
-    "whatwg-url": "^5.0.0",
-    "which": "^2.0.2",
-    "wide-align": "^1.1.5",
-    "wkx": "^0.5.0",
-    "word-wrap": "^1.2.3",
-    "wordwrap": "^1.0.0",
-    "wrappy": "^1.0.2",
-    "yallist": "^4.0.0"
-  },
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC"
-}
