+++ public/css/style.css
@@ -1,6 +1,5 @@
-
 header, .project:not(:last-child) {
-  border-bottom: 1px solid #c89292;
+  border-bottom: 1px solid #ccc;
 }
 
 .no-button {
