+++ views/homepage.handlebars
@@ -1,23 +1,23 @@
+<!-- Loop over each object in the projects array -->
+{{#each projects as |project| }}
+<!DOCTYPE html>
+<html lang="en">
 
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Book Worm</title>
   <link rel="stylesheet" href="css/style.css">
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bulma@0.9.4/css/bulma.min.css">
+  <div class="container is-fluid">
   <nav class="navbar" id="theNav" role="navigation" aria-label="main navigation">
     <div class="navbar-brand">
-      <div class="mainHeader">
+    <div class = "mainHeader">
         <h1>Book Worm</h1>
-      </div>
+    </div>
       <a role="button" class="navbar-burger" aria-label="menu" aria-expanded="false" data-target="navbarBasicExample">
         <span aria-hidden="true"></span>
         <span aria-hidden="true"></span>
         <span aria-hidden="true"></span>
       </a>
-    </div>
     <div id="navbarBasicExample" class="navbar-menu">
-      <div class="navbar-start" id="genre">
+      <div class="navbar-start" id="genre" >
         <a class="navbar-item">
           Home
         </a>
@@ -25,14 +25,14 @@
           Fantasy
         </a>
         <a class="navbar-item">
-          Sci-fi
-        </a>
-        <a class="navbar-item">
-          Non-Fiction
-        </a>
-        <a class="navbar-item">
-          Fiction
-        </a>
+           Sci-fi
+          </a>
+          <a class="navbar-item">
+            Non-Fiction
+          </a>
+          <a class="navbar-item">
+            Fiction
+          </a>
       </div>
       <div class="navbar-end" id="logIn">
         <div class="navbar-item">
@@ -47,20 +47,6 @@
         </div>
       </div>
     </div>
+    </div>
   </nav>
 </head>
-
-<body>
-  <!-- Loop over each object in the projects array -->
-  {{#each projects as |project| }}
-  <div class="container is-fluid">
-
-  </div>
-  {{/each}}
-  <script src="/js/script.js"></script>
-
-  <!-- Additional HTML content goes here -->
-
-</body>
-
-
