+++ views/layouts/main.handlebars
@@ -17,7 +17,6 @@
         <a href="/">Fund My Project!</a>
       </h1>
       <nav>
-        
         <!-- Conditionally render login or logout links -->
         {{#if logged_in}}
         <a href="/profile">profile</a> |
@@ -40,7 +39,6 @@
   {{#if logged_in}}
   <script src="/js/logout.js"></script>
   {{/if}}
-
 </body>
 
 </html>
